requirements.txt
flask>=0.12.2
gunicorn>=19.7.0
gevent>=1.4
pyyaml>=5.3.1
requests>=2.25.0
rdflib>=5.0.0
lxml>=4.6.2
bsddb3>=6.2.9
